fix(hint): nolazyredraw
When 'lazyredraw' is on, hydra hints don't always render correctly in
the cmdline. Specifically, the hint will render correctly when first
invoking the hydra, but any subsequent actions cause a block cursor to
appear at the end of the hint (and disappear from the active window).

This can be demonstrated (prior to this commit) by invoking nvim like,

  nvim -u minimal.lua +'call append(0, "1234")' +'goto 1'

with the following minimal.lua init file:

  vim.opt.lazyredraw = true
  vim.opt.wrap = false -- allow horizontal scrolling
  local hydra = require('hydra')
  hydra({
    name = 'scroll',
    mode = 'n',
    body = 'z',
    heads = { { 'l', 'zl' }, { 'h', 'zh' } },
    config = { hint = { type = 'cmdline' } },
  })

Type 'zl' to invoke the hydra, and the hint is drawn correctly. Type
either 'l' or 'h' to scroll again, and now the cursor is drawn to the
right of the hint in the echo area.